Ingest the Knesset rules of procedure into a model (#4)
Ingest the Knesset rules of procedure into a model

Read the latest version of the Knesset Rules of Procedure book
("Takanon") from the Hebrew WikiSource and save each clause as
a model item. If a clause is already saved, it is updated.

The clause text is a semi-structured JSON document capturing the
breakdown into subclauses, sub-subclauses, etc.

Amendments to clauses are tracked by scraping all available
WikiSource revisions in order.